refactor: security 관련 config수정
