+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/jwt/JwtFilterConfig.java
@@ -1,11 +1,11 @@
-package sellyourunhappiness.global.config;
+package sellyourunhappiness.global.config.jwt;
 
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class FilterConfig {
+public class JwtFilterConfig {
 
 	@Bean
 	public FilterRegistrationBean<JwtTokenFilter> jwtTokenFilterRegistrationBean(JwtTokenFilter jwtTokenFilter) {
