+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/security/handler/CustomOAuth2LoginFailureHandler.java
@@ -1,4 +1,4 @@
-package sellyourunhappiness.core.security.handler;
+package sellyourunhappiness.global.config.security.handler;
 
 import java.io.IOException;
 
@@ -15,10 +15,10 @@ import lombok.extern.slf4j.Slf4j;
 @Component
 public class CustomOAuth2LoginFailureHandler implements AuthenticationFailureHandler {
 	@Override
-	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws
+	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException exception) throws
 		ServletException, IOException {
 		response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-		response.getWriter().write("소셜 로그인 실패! 서버 로그를 확인해주세요.");
-		log.info("소셜 로그인에 실패했습니다. 에러 메시지 : {}", exception.getMessage());
+		response.getWriter().write("소셜 로그인 실패, 서버 로그를 확인해주세요.");
 	}
 }
